+++ src/clj_cbor/core.clj
@@ -81,18 +81,25 @@
 (defn- data-output-stream
   "Coerce the argument to a `DataOutputStream`."
   ^DataOutputStream
-  [input]
-  (if (instance? DataOutputStream input)
-    input
-    (DataOutputStream. (io/output-stream input))))
+  [output]
+  (condp instance? output
+    DataOutputStream
+    output
+
+    OutputStream
+    (DataOutputStream. output)
+
+    (throw (IllegalArgumentException.
+             (str "Cannot coerce argument to an OutputStream: "
+                  (pr-str output))))))
 
 
 (defn encode
   "Encode a single value as CBOR data.
 
-  In the full argument form, this writes a value to an output stream and
-  returns the number of bytes written. If output is omitted, the function
-  returns a byte array instead. Uses the `default-codec` if none is provided."
+  Writes the value bytes to the provided output stream, or returns the value
+  as a byte array if no output is given. The `default-codec` is used to encode
+  the value if none is provided."
   ([value]
    (encode default-codec value))
   ([encoder value]
@@ -100,7 +107,7 @@
      (with-open [output (data-output-stream buffer)]
        (encode encoder output value))
      (.toByteArray buffer)))
-  ([encoder ^OutputStream output value]
+  ([encoder output value]
    (let [data-output (data-output-stream output)]
      (codec/write-value encoder data-output value))))
 
@@ -109,9 +116,9 @@
   "Encode a sequence of values as CBOR data. This eagerly consumes the
   input sequence.
 
-  In the full argument form, this writes a value to an output stream and
-  returns the number of bytes written. If output is omitted, the function
-  returns a byte array instead. Uses the `default-codec` if none is provided."
+  Writes the value bytes to the provided output stream, or returns the value
+  as a byte array if no output is given. The `default-codec` is used to encode
+  the value if none is provided."
   ([values]
    (encode-seq default-codec values))
   ([encoder values]
@@ -119,7 +126,7 @@
      (with-open [output (data-output-stream buffer)]
        (encode-seq encoder output values))
      (.toByteArray buffer)))
-  ([encoder ^OutputStream output values]
+  ([encoder output values]
    (let [data-output (data-output-stream output)]
      (transduce (map (partial encode encoder data-output)) + 0 values))))
 
@@ -130,8 +137,13 @@
 (defn- data-input-stream
   "Coerce the argument to a `DataInputStream`."
   [input]
-  (if (instance? DataInputStream input)
+  (condp instance? input
+    DataInputStream
     input
+
+    InputStream
+    (DataInputStream. input)
+
     (DataInputStream. (io/input-stream input))))
 
 
@@ -161,9 +173,14 @@
 (defn decode
   "Decode a single CBOR value from the input.
 
-  The input may be a byte array or coercible to an `input-stream`. Uses the
-  `default-codec` if none is provided. If at the end of the input, returns nil
-  or `eof-guard` if provided."
+  This uses the given codec or the `default-codec` if none is provided. If at
+  the end of the input, this returns `eof-guard` or nil.
+
+  The input must be an input stream or something coercible to one like a file
+  or byte array. Note that coercion will produce a `BufferedInputStream` if the
+  argument is not already a stream, so repeated reads will probably not behave
+  as expected! If you need incremental parsing, make sure you pass in something
+  that is already an `InputStream`."
   ([input]
    (decode default-codec input))
   ([decoder input]
@@ -179,9 +196,12 @@
 (defn decode-seq
   "Decode a sequence of CBOR values from the input.
 
-  The input may be a byte array or coercible to an `input-stream`. Uses the
-  `default-codec` if none is provided. This returns a lazy sequence, so take
-  care that the input stream is not closed before the entries are realized."
+  This uses the given codec or the `default-codec` if none is provided. The
+  returned sequence is lazy, so take care that the input stream is not closed
+  before the entries are realized.
+
+  The input must be an input stream or something coercible to one - see
+  `decode` for usage notes."
   ([input]
    (decode-seq default-codec input))
   ([decoder input]