+++ CHANGELOG.md
@@ -5,9 +5,31 @@ All notable changes to this project will be documented in this file, which
 follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [Unreleased]
 
 ...
+
+
+## [0.7.2] - 2019-06-28
+
+### Added
+- Add [clj-async-profiler](https://github.com/clojure-goes-fast/clj-async-profiler)
+  support in the benchmarking code to produce flame graphs for finding hotspots.
+
+### Removed
+- Codecs no longer support a configurable set tag; it is now fixed at 258
+  matching the IANA registry.
+
+### Fixed
+- `decode` will not try to coerce inputs which are already input streams
+  anymore. This fixes incremental/lazy decoding behavior.
+  [#13](https://github.com/greglook/clj-cbor/issues/13)
+- `encode` will not try to coerce outputs to an output stream anymore. Now an
+  exception is thrown when a bad argument is given. This prevents lost writes
+  when the final buffer is not flushed correctly.
+  [#14](https://github.com/greglook/clj-cbor/pull/14)
+
 
 ## [0.7.1] - 2019-01-10
 
@@ -31,10 +53,12 @@ decoding time from 8.454 ms to 4.089 ms, or about twice as fast!
 - A tagged-value major type with a streaming info code now results in a
   `::codec/illegal-stream` error.
 
+
 ## [0.6.0] - 2017-12-23
 
 ### Changed
 - Upgrade to Clojure 1.9.0.
+
 
 ## [0.5.0] - 2017-11-08
 
@@ -79,6 +103,7 @@ using a custom read handler:
 - `clj-cbor.core/spit-all` similarly writes a sequence of values to an output
   file like repeated calls to `spit` with `:append true`.
 
+
 ## [0.4.1] - 2017-05-17
 
 ### Fixed
@@ -89,10 +114,12 @@ using a custom read handler:
   [#4](https://github.com/greglook/clj-cbor/issues/4)
   [#5](https://github.com/greglook/clj-cbor/pull/5)
 
+
 ## [0.4.0] - 2017-03-14
 
 ### Added
 - Implemented canonical mode sorting of map keys and set entries.
+
 
 ## [0.3.0] - 2016-01-05
 
@@ -114,6 +141,7 @@ using a custom read handler:
 ### Fixed
 - The `decode` function now properly raises an error when the input ends
   mid-value rather than at a top-level value boundary.
+
 
 ## [0.2.0] - 2016-12-28
 
@@ -137,11 +165,14 @@ This release includes 100% test coverage!
 ### Fixed
 - Generally tighted up namespaces and reduced linkage where possible.
 
+
 ## [0.1.0] - 2016-12-23
 
 Initial project release.
 
-[Unreleased]: https://github.com/greglook/clj-cbor/compare/0.7.1...HEAD
+
+[Unreleased]: https://github.com/greglook/clj-cbor/compare/0.7.2...HEAD
+[0.7.2]: https://github.com/greglook/clj-cbor/compare/0.7.1...0.7.2
 [0.7.1]: https://github.com/greglook/clj-cbor/compare/0.6.0...0.7.1
 [0.6.0]: https://github.com/greglook/clj-cbor/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/greglook/clj-cbor/compare/0.4.1...0.5.0